rename file to occ-loaded-config-converter.ts
