+++ projects/core/src/occ/config-loader/index.ts
@@ -1,3 +1,3 @@
 export { OccConfigLoaderModule } from './occ-config-loader.module';
 export * from './occ-loaded-config';
-export * from './occ-loaded-config.converter';
+export * from './occ-loaded-config-converter';
