+++ projects/core/src/occ/config-loader/occ-loaded-config-converter.spec.ts
@@ -7,7 +7,7 @@ import {
 import { Occ } from '../occ-models';
 import { JavaRegExpConverter } from './java-reg-exp-converter';
 import { OccLoadedConfig } from './occ-loaded-config';
-import { OccLoadedConfigConverter } from './occ-loaded-config.converter';
+import { OccLoadedConfigConverter } from './occ-loaded-config-converter';
 
 describe(`OccLoadedConfigConverter`, () => {
   let converter: OccLoadedConfigConverter;
